/!\ API change: introduce Question component and port questions API /!\
/!\ this breaks the children API /!\

instead of passing just the label and constructing all the component in a
loop, we create a Question component that you can use instead of the <span/>
if you only want the previous API, or you can implement your own that just
needs to implement an onChange callback whenever the data is ready to be
added to values.

this allows to implement maps, typeahead, or any other kind of custom
components. we should probably ship some of those so that we have pretty and
useable stuff right away.