Reimplement Animations (still needs fixing)
we completly rework how animation work, introducing a `.current`, `.next`
and `.prev` system of classNames, we also introduce a `Number` component
(that we export so you can re-use it in other parts of your UI) that uses
the same CSS as the questions roll to handle the counter.

we will animate any `<label/>` we find, so if you want something to be
roll-animated in your Component just wrap it in a `<label/>` (maybe we need
to className it ?), all other components will just disappear on transition.